requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
PyGithub>=2.1.1
